Simplify CNB local review tooling
