+++ climate-advisor/scripts/cnb_research/review.html
@@ -6,6 +6,8 @@
     <meta name="description" content="Review CNB funded-project research and similar-project search results." />
     <title>CNB Local Review</title>
     <link rel="stylesheet" href="review.css" />
+    <script src="review-support.js" defer></script>
+    <script src="review-renderer.js" defer></script>
     <script src="review.js" defer></script>
   </head>
   <body>
